llama-2-local/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.5
brotlipy==0.7.0
build==0.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
comm==0.1.3
cryptography==39.0.1
ctransformers==0.2.5
dataclasses-json==0.5.13
debugpy==1.6.7
decorator==5.1.1
diskcache==5.6.1
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.100.0
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.16.4
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.225
langchainplus-sdk==0.0.20
llama-cpp-python==0.1.77
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.0.1
platformdirs==3.9.1
priority==2.0.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==23.0.0
pypdf==3.8.1
pyproject_hooks==1.0.0
PySocks==1.7.1
python-box==7.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.0
quart==0.18.4
quart-cors==0.6.0
regex==2023.6.3
requests==2.29.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==66.0.0
singlestoredb==0.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparams==5.1.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.7.1
typing-inspect==0.9.0
urllib3==1.26.15
uvicorn==0.23.1
wcwidth==0.2.6
Werkzeug==2.3.4
wheel==0.38.4
wsproto==1.2.0
yarl==1.9.2
